Updating states , task and parallel
